Improve reliability of uiBokehMgr
Summary:
Principal changes:
- Replace use of bokeh serve with embedded stand-alone bokeh servers in the main.py of  bokeh_crossplot and bokeh_machine_learning
 Python apps
- Use a uiBokehServiceMgr to communicate with the individual bokeh servers to retrieve network connection details and remove
the need to parse log files for this information
- Noticed and sort of fixed a potential race condition with allocation of port numbers for the bokeh servers. dgbServerInfo gets
a port number in its constructor but the port is not actually claimed until bokeh starts so it's possible for the same port to be
 picked up by other dgbServerInfo instances and used to attempt to start other bokeh servers. I've moved the creation of
the dgbServerInfo instance to as close as possible to where the bokeh server script is executed to reduce the risk.
Ultimate solution would be to let each bokeh server get a port number when it starts.
- I've also moved the training related messages and callbacks from the uidGBServiceMgr into the uiBokehServiceMgr
- I've commented out the methods/callbacks used for log handling I think are now redundant
Fixes OD-2528 and OD-2495

Test Plan: It's working under Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13152