dgb: solved various issues in ML train tab
Fixed various issues related to the callbacks of that tab. Now the output name check is done when pressing
the run button of the Python user-interface, otherwise typing a new string directly in the combo-box would create issues.
Slightly reworked the layout the make it more intuitive.
Also made the uiBokeh::Setup more standard, with support for default sizes.