Machine learning apply can now predict on CPU in GPU environments
Static fix for OD-2549:
The user interface now has a checkbox to run the prediction on the CPU at request.
Also added a user interface to set the batch size used for the prediction,
although this later is limited to 1 for the img2img workflow. Can be used to
improve the efficiency and memory consumption of the prediction.
Finally fixed a minor crash, if all log predictions fail, that prevented to error
message to be displayed.