Use raster based shading in Bokeh logplot - show depths as positive numbers - interpolate logs on input to common depth index
Summary:
This incorporates into master the new work done by Paul to do log shading using an image.
To get this to work with the F3 Demo project I had to build a log interpolation step
into the log input code. It appears the depth index values being read have a rounding error
such that the depth index of logs starting at different depths get out of sync and if simply
merged into a Pandas dataframe data gets dropped and NaN's get inserted and the log won't
plot as a continuous line. The change fixes that but adds a little bit of time to the loading
of logs.
Displaying the shading using an image seems to work OK. I couldn't see the display issue
described by Paul. Could it be a browser issue? Maybe the webkit in Qt5.14.2 used by master is
better than the stand-a-lone browser he was testing on.
A few issues I see with the app are:
- when displaying more than 2 logs in a track there is no axis for the the 3rd or later logs
- it isn't obvious which log curve belongs to which axis - maybe the axis color should match
the log color
- only 2 logs can be displayed in a track when shading between logs
- when shading between logs the axis for the 2nd log no longer has any meaning

I couldn't see a quick fix for these issues without substatnially changing the implementation.
Fixes OD-2652

Test Plan: Tetsted on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13303