Revamp of Bokeh log plotting
Summary:
I've completely rewritten the Bokeh log plotting to make it easier to use, maintain
and extend and hopefully faster as well.  I've avoided the use of a Pandas dataframe
so no log resampling is required. Each log is loaded with its depth index only when
it is first needed. I'm using the odpy.wellman to get well information, log data and
markers so the commandline really only needs the well id and possibly a list of logs
to display. At the moment the log selection is hard wired. It is possible to have
any number of logs displayed each with individual axis ranges.

This is a work in progress. I need to:
- Add a UI tab so the user can set the displayed log range and display style
- Add a multiselect box below each track so the user can change the logs displayed
- Shading (need to test if using the built-in Bokeh shading can be used, otherwise
I'll add Pauls background image idea)
- Add  buttons to add/remove additional tracks from the display
- I'm assuming the inital set of logs displayed is provided on the commandline

It can be started from OpendTect and will honour the well selected but the logs displayed
 are hardwired. You can also execute bokeh_logplot.py in Spyder (started from the OpendTect
 toolbar) and it will read the well data according to the code in the main function from
the current project (well selected assumes the F3 Demo project) and display in the default
 browser. For this to work make sure "Execute in an external system terminal" under Console
and "Interact with the Python console after execution" under External system terminal are
set in the Run configuration per file.
This is what it looks like:
{F2303}

Test Plan: Tested on Linux

Reviewers: nanne, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13321