Add UI and multitrack display to new bokeh logplot
Summary:
This version of the new Bokeh logplot comes with a UI to:
- control log appearance (color, line thickness and dashing)
- control log plotting scales
- select logs and markers to display
It is also possible to display multiple tracks each with its own set of display parameters.
odpy.wellman is used for reading log data, no resampling required. Here's a video of it
in action:
{F2305}

What hasn't been done:
- logarithmic log scale
- UI setting of the track and depth grid
- shading
- any sort of linked zooming/panning of the tracks

Test Plan: Tested on Linux

Reviewers: nanne, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13330