Bokeh logplot usability changes
Summary:
Fixes OD-2923 Well Table Management- Log Plot (Bokeh): suggestion to re-organize the tables
Wasn't able to introduce tooltips for the buttons used to duplicate/remove log tracks, Bokeh
doesn't provide this feature at the moment.
Changed the UI tab order as requested. Now if only a single log is passed on the command
line the UI will open of the Data tab so more logs to display can be selected
Also added a checkbox that determines if a track is scrolled or panned in sync with the
master track (the leftmost track in the display)
The Bokeh tools are now hidden for tracks that are in sync. An unfortunate side effect
is that zooming and panning for "synced" tracks is only active if the mouse cursor is
 in the master track

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D13590