Added periodic saving of keras model while training on epoch ended
Should allow to have a readable model if the training crashes on
epoch 14/20, the last saved model will get written to the OpendTect
database.

Temporary models gets saved in raw keras format in the Survey/Proc
directory under a temporary name.
Saving may not strictly be done at the end of each epoch, there is a limit
set to 5 minutes between consecutive saving.