Tentative fix for invalid argument error with tensorflow
InvalidArgumentError is issued sometimes while training a model, or
predicting from a model.
Tensorflow removed the possibility to use a Sequence generator object for
providing the validation data. Thus changed the code to be able to provide
back a set of numpy arrays.
This code will provide compatibility for the current tensorflow version we are using (2.1),
but also newer versions, which introduce a separate batch size for the validation data.