Added meta-data based pre-trained machine learning models [devel]
Summary:
First step for supporting downloadable pre-trained models. The minimum
information is written in a json file, read by OpendTect. When parsed by OpendTect,
the path should be used to trigger the download of the file and its import
into the target (current) project (not part of this revision yet).
The existing pre-trained model is modified, as both the json and hdf5 got merged.

Test Plan:
hdf5 and json-based models are listed in the control center top dialog.
Fault predictor can be applied, the two others issue a TODO error.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D13933