Ensure C++ ui settings are sent to Bokeh at startup
Summary:
Fixes OD-3336 Transfer training does not make sense
The Bokeh UI isn't started until an Example file is provided. Previously if the user changes the
Training Type before selecting an Example File the message is lost (no Bokeh UI to receive it).
Proposed change ensures the Transfer Type setting is sent to the Bokeh UI after it is up and running.
Also fixed an issue with the Bokeh UI failing on restart when the TrainingType is "transfer".

Test Plan: Tested on Linux

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14181