Improvements in Bokeh training app [devel]
To facilitate its development/maintenance, made it
possible to run the app without a running OpendTect
Machine Learning control center.

Since no bokeh server manager is running on the C++ side,
no communication occurs and all required data must be provided
up front, like the to-be-trained model dataset name, which must also
already be registered in the OpendTect database with the appropriate type beforehand.

Also added a couple of command line arguments defaults, to minimize the amount of arguments
that have to be provided. In debug mode the script minimum is the path to the example file
and output dataset (model) name.