Fixed training using SVM regression models
LinearSVM class does not have either kernel or degree arguments,
as it is a sub implementation of SVM class limited to the linear case.

Changed some defaults, to use LinearSVC or LinearSVM classes for the linear
kernel, as it is much faster and supports larger training sets, besides
SVC and SVM classes both breakdown with large number of input attributes.

[Jira: OD-3241]