Enable apply of 2D crossline models
Summary:
Changes allow 2D crossline models (eg with shape 64_1_128) to be applied to
2D data and to 3D data (in either inline, crossline or average mode).
It was only necessary to override the model shape to get this working and then apply
the model just like a 2D inline model (eg with shape 1_64_128). Some of the logic was simplified
as a result.

Test Plan:
Tested on Linux
Img2Img
- 2D model apply on 3D: crossline model in inline, crossline and average: ok
- 2D model apply on 2D: crossline model: ok

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15562