Added input to the slider field and fixed some incompatibility issues
Summary:
[OD-6264] Added an input field that synchronizes with the slider field.
Fixed an error and logic related to reading the ML Platform.
Updated the arguments used in torch.amp.autocast.

Test Plan: In the Machine Learning Plugins, train a model using a Torch-based architecture. Interact with both the slider and input field. Verify that changes in the slider are reflected in the input field and vice versa.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17907