Fixed Seismic Classification Apply Multiple Outputs
Summary: [OD-6309] Previously, there was an error message when applying the seismic classification model with multiple outputs. It turned out that the number of outputs did not match between the UI input and the Python side. The bug was fixed by changing the shape of prediction in dgbtorch seismic classification.

Test Plan: Open the Machine Learning Plugins and go to the Seismic Classification workflow. Apply the Torch model and select multiple outputs. Make sure the program runs correctly.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com/D17932